+++ src/libs/bemedev/globals/types.ts
@@ -1,0 +1,4 @@
+export * from '../features/arrays/types';
+export * from '../features/common/types';
+export * from '../features/functions/types';
+export * from '../features/objects/types';